Fix Call Library Function Node crashing LabVIEW
The writer "handle", which is a `uintptr_t` should be passed by value NOT by reference (Pointer by Value). LabVIEW automatically dereferences the pointer to an address in memory as a value, so passing by reference was causing LabVIEW to crash. I should have known this by looking at the `Destroy` VI for the Writer class. Anyways, LabVIEW no longer crashes with Begin and End tests. I am not even sure that the Initialize and Shutdown were causing problems and had to be removed. I do like it without them as it is one less thing needed to use the library.

git-svn-id: svn://home-office.fios-router.home/LabPack/trunk@18 20bef0b3-2b52-1341-a172-4f937900580d