Change Shared Library path connector location in VIs
The Shared Library control, which is optional, is moved from the top to the bottom of the From and To Msgpack VIs. This is to be consistent with the SQLite library by DJDPowell and the JSON serialization VIs from JKI. The location still needs to be moved for the Utility VIs, the Writer class, and the Reader class.

git-svn-id: svn://home-office.fios-router.home/LabPack/trunk@117 20bef0b3-2b52-1341-a172-4f937900580d