Change on-disk location for HTML files to sub-folder for help
The HTML Help files are moved to a `html` sub-folder to avoid cluttering up the parent directory for the Help source with a bunch of HTML files. There were already sub-folders for CSS, JavaScript, and Images. It makes sense to have a HTML only folder. Then, only HTML Help Workshop files are present in the parent directory for the Help source code.

git-svn-id: svn://localhost/LabPack/trunk@172 20bef0b3-2b52-1341-a172-4f937900580d