Remove Caraya license from Readme
The Caraya license information is removed from the HTML Help Readme file because Caraya VIs are NOT distributed with "release" package of the library. Caraya VIs are only used in the tests as part of the project/source code, so the Caraya license is included in the source code distribution, not the deployment package (VI Package Manager).

git-svn-id: svn://home-office.fios-router.home/LabPack/trunk@217 20bef0b3-2b52-1341-a172-4f937900580d