Change CHM output location
The CHM file needs to be in the `src` hierarchy so it can be included in the package distribution. The `.chm` file type is added to the ignore list instead. The help project is now compiled to a .chm file in the Help source folder but it is excluded version control.

The package build specification is also updated to excluded all non-chm files and place files in appropriate locations on install.

git-svn-id: svn://home-office.fios-router.home/LabPack/trunk@232 20bef0b3-2b52-1341-a172-4f937900580d