Fix VIP build for LabVIEW 2015
The older version of LabVIEW that can be supported is 2015. There are VIs for data type information that are only available after LabVIEW 2014.

Note, the automated build does not currently work with Labricator because Caraya v1.0 has a regression where it brings in the user.lib OpenG dependencies instead of including them as internal dependencies.